Review Add-Ons by "known" concrete5 developers
Make a video where [x] Add-Ons will be installed and show its strenght and weaknesses. Maybe always do [x] "old"Add-Ons and [x] recently added one? The reviewers could be core developers or people from the community and should be more than just 1 to make it a "real" thing which people WANT to watch.

Having a textual version to read (instead of watch) may be pretty useful too (for when people don't have the time or can't watch videos because on their mobile device/network).